+++ fractal_server/app/models/v2/project.py
@@ -14,8 +14,10 @@ from fractal_server.utils import get_timestamp
 class ProjectV2(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     name: str
+
+    # TODO-2.17.1: make `resource_id` not nullable
     resource_id: int | None = Field(
-        foreign_key="resource.id", default=None, ondelete="SET NULL"
+        foreign_key="resource.id", default=None, ondelete="RESTRICT"
     )
     timestamp_created: datetime = Field(
         default_factory=get_timestamp,