+++ fractal_server/app/routes/auth/oauth.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter
 from httpx_oauth.clients.github import GitHubOAuth2
 from httpx_oauth.clients.google import GoogleOAuth2
 from httpx_oauth.clients.openid import OpenID
+from httpx_oauth.clients.openid import OpenIDConfigurationError
 
 from . import cookie_backend
 from . import fastapi_users
@@ -26,11 +27,21 @@ def _create_client_google(cfg: OAuthSettings) -> GoogleOAuth2:
 
 
 def _create_client_oidc(cfg: OAuthSettings) -> OpenID:
-    return OpenID(
-        client_id=cfg.OAUTH_CLIENT_ID.get_secret_value(),
-        client_secret=cfg.OAUTH_CLIENT_SECRET.get_secret_value(),
-        openid_configuration_endpoint=cfg.OAUTH_OIDC_CONFIG_ENDPOINT.get_secret_value(),  # noqa
-    )
+    try:
+        open_id = OpenID(
+            client_id=cfg.OAUTH_CLIENT_ID.get_secret_value(),
+            client_secret=cfg.OAUTH_CLIENT_SECRET.get_secret_value(),
+            openid_configuration_endpoint=cfg.OAUTH_OIDC_CONFIG_ENDPOINT.get_secret_value(),  # noqa
+        )
+    except OpenIDConfigurationError as e:
+        OAUTH_OIDC_CONFIG_ENDPOINT = (
+            cfg.OAUTH_OIDC_CONFIG_ENDPOINT.get_secret_value()
+        )
+        raise RuntimeError(
+            f"Cannot initialize OpenID client. Original error: '{e}'. "
+            f"Hint: is {OAUTH_OIDC_CONFIG_ENDPOINT=} reachable?"
+        )
+    return open_id
 
 
 def get_oauth_router() -> APIRouter | None:
@@ -44,7 +55,6 @@ def get_oauth_router() -> APIRouter | None:
         return None
 
     client_name = oauth_settings.OAUTH_CLIENT_NAME
-
     if client_name == "google":
         client = _create_client_google(oauth_settings)
     elif client_name == "github":