+++ fractal_server/app/routes/auth/users.py
@@ -15,9 +15,8 @@ from . import current_superuser_act
 from ...db import get_async_db
 from ...schemas.user import UserRead
 from ...schemas.user import UserUpdate
-from ._aux_auth import _get_default_usergroup_id
+from ._aux_auth import _get_default_usergroup_id_or_none
 from ._aux_auth import _get_single_user_with_groups
-from ._aux_auth import FRACTAL_DEFAULT_GROUP_NAME
 from fractal_server.app.models import LinkUserGroup
 from fractal_server.app.models import UserGroup
 from fractal_server.app.models import UserOAuth
@@ -26,7 +25,10 @@ from fractal_server.app.routes.auth._aux_auth import _user_or_404
 from fractal_server.app.schemas.user import UserUpdateGroups
 from fractal_server.app.security import get_user_manager
 from fractal_server.app.security import UserManager
+from fractal_server.config import get_settings
 from fractal_server.logger import set_logger
+from fractal_server.syringe import Inject
+
 
 router_users = APIRouter()
 
@@ -146,6 +148,7 @@ async def set_user_groups(
     superuser: UserOAuth = Depends(current_superuser_act),
     db: AsyncSession = Depends(get_async_db),
 ) -> UserRead:
+    settings = Inject(get_settings)
     # Preliminary check that all objects exist in the db
     user = await _user_or_404(user_id=user_id, db=db)
     target_group_ids = user_update.group_ids
@@ -161,13 +164,16 @@ async def set_user_groups(
         )
 
     # Check that default group is not being removed
-    default_group_id = await _get_default_usergroup_id(db=db)
-    if default_group_id not in target_group_ids:
+    default_group_id_or_none = await _get_default_usergroup_id_or_none(db=db)
+    if (
+        default_group_id_or_none is not None
+        and default_group_id_or_none not in target_group_ids
+    ):
         raise HTTPException(
             status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
             detail=(
                 f"Cannot remove user from "
-                f"'{FRACTAL_DEFAULT_GROUP_NAME}' group.",
+                f"'{settings.FRACTAL_DEFAULT_GROUP_NAME}' group.",
             ),
         )