+++ tests/v2/test_03_api/admin/test_admin_resource.py
@@ -166,7 +166,9 @@ async def test_resource_api(
         profile = res.json()
         res = await client.delete(f"/admin/v2/resource/{resource_id}/")
         assert res.status_code == 422
-        assert "it's associated with 1 Profiles" in str(res.json()["detail"])
+        assert (
+            'key constraint "profile_resource_id_fkey"' in res.json()["detail"]
+        )
         res = await client.delete(f"/admin/v2/profile/{profile['id']}/")
         assert res.status_code == 204
 
@@ -176,7 +178,10 @@ async def test_resource_api(
         await db.refresh(project)
         res = await client.delete(f"/admin/v2/resource/{resource_id}/")
         assert res.status_code == 422
-        assert "it's associated with 1 Project" in str(res.json()["detail"])
+        assert (
+            'key constraint "projectv2_resource_id_fkey"'
+            in res.json()["detail"]
+        )
         await db.delete(project)
         await db.commit()
 
@@ -191,8 +196,9 @@ async def test_resource_api(
         await db.refresh(task_group)
         res = await client.delete(f"/admin/v2/resource/{resource_id}/")
         assert res.status_code == 422
-        assert "it's associated with 1 TaskGroupV2" in str(
-            res.json()["detail"]
+        assert (
+            'key constraint "taskgroupv2_resource_id_fkey"'
+            in res.json()["detail"]
         )
         await db.delete(task_group)
         await db.commit()